Honor header whitelist for Powerbox HTTP APIs.
Fixes #2944
